CreateSusu: Validations modified.
