+++ src/Domain/Susu/BizSusu/Actions/Create/BizSusuCreateSusuAmountAction.php
@@ -7,8 +7,8 @@ namespace Domain\Susu\BizSusu\Actions\Create;
 use Domain\Shared\Action\General\CreateSusuValidationAction;
 use Domain\Shared\Action\Session\SessionInputUpdateAction;
 use Domain\Shared\Menus\General\CreateSusuValidationMenu;
+use Domain\Shared\Menus\General\GeneralMenu;
 use Domain\Shared\Models\Session\Session;
-use Domain\Susu\BizSusu\Menus\Create\BizSusuCreateMenu;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
 final class BizSusuCreateSusuAmountAction
@@ -30,6 +30,6 @@ final class BizSusuCreateSusuAmountAction
         SessionInputUpdateAction::updateUserInputs(session: $session, user_input: ['susu_amount' => $session_data->user_input]);
 
         // Return the initialDepositMenu
-        return BizSusuCreateMenu::initialDepositMenu(session: $session);
+        return GeneralMenu::initialDepositMenu(session: $session);
     }
 }
