+++ src/Domain/Susu/FlexySusu/Menus/Create/FlexySusuCreateMenu.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace Domain\Susu\FlexySusu\Menus\Create;
 
-use App\Common\LinkedWallets;
 use App\Common\ResponseBuilder;
-use App\Common\SusuResources;
 use Domain\Shared\Models\Session\Session;
 use Domain\Susu\Shared\Actions\Common\GetSusuFrequenciesAction;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -20,56 +18,6 @@ final class FlexySusuCreateMenu
 
         return ResponseBuilder::ussdResourcesResponseBuilder(
             message: 'Enter account name',
-            session_id: $session->session_id,
-        );
-    }
-
-    public static function susuAmount($session): JsonResponse
-    {
-        // Return the debitFrom
-        return ResponseBuilder::ussdResourcesResponseBuilder(
-            message: 'Enter susu amount',
-            session_id: $session->session_id,
-        );
-    }
-
-    public static function initialDeposit($session): JsonResponse
-    {
-        return ResponseBuilder::ussdResourcesResponseBuilder(
-            message: 'Start with (amount)',
-            session_id: $session->session_id,
-        );
-    }
-
-    public static function frequencyMenu($session): JsonResponse
-    {
-        // Get the frequencies from the session->user_data
-        $frequencies = json_decode($session->user_data, associative: true);
-
-        return ResponseBuilder::ussdResourcesResponseBuilder(
-            message: "Choose frequency\n".SusuResources::formatFrequenciesForMenu(start_dates: $frequencies['frequencies']),
-            session_id: $session->session_id,
-        );
-    }
-
-    public static function invalidFrequencyMenu($session): JsonResponse
-    {
-        // Get the frequencies from the session->user_data
-        $frequencies = json_decode($session->user_data, associative: true);
-
-        return ResponseBuilder::ussdResourcesResponseBuilder(
-            message: "Invalid frequency\n".SusuResources::formatFrequenciesForMenu(start_dates: $frequencies['frequencies']),
-            session_id: $session->session_id,
-        );
-    }
-
-    public static function linkedWalletMenu($session): JsonResponse
-    {
-        // Get the linked_wallets from the session->user_data
-        $linked_wallets = json_decode($session->user_data, associative: true);
-
-        return ResponseBuilder::ussdResourcesResponseBuilder(
-            message: "Choose wallet\n".LinkedWallets::formatLinkedWalletForMenu(linked_wallets: $linked_wallets['linked_wallets']),
             session_id: $session->session_id,
         );
     }
