+++ src/Domain/Susu/GoalGetterSusu/Actions/Create/GoalGetterSusuCreateDurationAction.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Domain\Susu\GoalGetterSusu\Actions\Create;
 
 use Domain\Shared\Action\Session\SessionInputUpdateAction;
+use Domain\Shared\Menus\General\GeneralMenu;
 use Domain\Shared\Models\Session\Session;
-use Domain\Susu\GoalGetterSusu\Menus\Create\GoalGetterSusuCreateMenu;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
 final class GoalGetterSusuCreateDurationAction
@@ -18,7 +18,7 @@ final class GoalGetterSusuCreateDurationAction
 
         // Validate the user_input (susu_amount)
         return match (true) {
-            ! array_key_exists(key: $session_data->user_input, array: $duration) => GoalGetterSusuCreateMenu::invalidDurationMenu(session: $session),
+            ! array_key_exists(key: $session_data->user_input, array: $duration) => GeneralMenu::invalidDurationMenu(session: $session),
 
             default => self::durationStore(session: $session, session_data: $session_data, duration: $duration)
         };
@@ -30,6 +30,6 @@ final class GoalGetterSusuCreateDurationAction
         SessionInputUpdateAction::updateUserInputs(session: $session, user_input: ['duration' => $duration[$session_data->user_input]['code']]);
 
         // Return the enterSusuAmountMenu
-        return GoalGetterSusuCreateMenu::startDateMenu(session: $session);
+        return GeneralMenu::startDateMenu(session: $session);
     }
 }
