+++ src/Domain/Susu/GoalGetterSusu/Actions/Create/GoalGetterSusuCreateTargetAmountAction.php
@@ -7,8 +7,8 @@ namespace Domain\Susu\GoalGetterSusu\Actions\Create;
 use Domain\Shared\Action\General\CreateSusuValidationAction;
 use Domain\Shared\Action\Session\SessionInputUpdateAction;
 use Domain\Shared\Menus\General\CreateSusuValidationMenu;
+use Domain\Shared\Menus\General\GeneralMenu;
 use Domain\Shared\Models\Session\Session;
-use Domain\Susu\GoalGetterSusu\Menus\Create\GoalGetterSusuCreateMenu;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
 final class GoalGetterSusuCreateTargetAmountAction
@@ -30,6 +30,6 @@ final class GoalGetterSusuCreateTargetAmountAction
         SessionInputUpdateAction::updateUserInputs(session: $session, user_input: ['target_amount' => $session_data->user_input]);
 
         // Return the initialDepositMenu
-        return GoalGetterSusuCreateMenu::initialDepositMenu(session: $session);
+        return GeneralMenu::initialDepositMenu(session: $session);
     }
 }
