+++ src/Domain/Susu/PersonalSusu/Actions/Create/PersonalSusuCreateAccountNameAction.php
@@ -7,8 +7,8 @@ namespace Domain\Susu\PersonalSusu\Actions\Create;
 use Domain\Shared\Action\General\CreateSusuValidationAction;
 use Domain\Shared\Action\Session\SessionInputUpdateAction;
 use Domain\Shared\Menus\General\CreateSusuValidationMenu;
+use Domain\Shared\Menus\General\GeneralMenu;
 use Domain\Shared\Models\Session\Session;
-use Domain\Susu\PersonalSusu\Menus\Create\PersonalSusuCreateMenu;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
 final class PersonalSusuCreateAccountNameAction
@@ -17,7 +17,7 @@ final class PersonalSusuCreateAccountNameAction
     {
         // Validate the user_input (account_name)
         return match (true) {
-            CreateSusuValidationAction::accountNameLength($session_data->user_input) === false => CreateSusuValidationMenu::accountNameLengthMenu(session: $session),
+            CreateSusuValidationAction::accountNameLengthValid($session_data->user_input) === false => CreateSusuValidationMenu::accountNameLengthMenu(session: $session),
 
             default => self::accountNameStore(session: $session, session_data: $session_data)
         };
@@ -29,6 +29,6 @@ final class PersonalSusuCreateAccountNameAction
         SessionInputUpdateAction::updateUserInputs(session: $session, user_input: ['account_name' => $session_data->user_input]);
 
         // Return the susuAmountMenu
-        return PersonalSusuCreateMenu::susuAmountMenu(session: $session);
+        return GeneralMenu::susuAmountMenu(session: $session);
     }
 }
